+++ src/types/UserProjectData.ts
@@ -1,5 +1,11 @@
 import { ProjectResource } from "./ProjectResource";
 
+export enum DiscoveryTrigger {
+    ProjectUpdate = "ProjectCreation",
+    UserManual = "UserManual",
+    AutomaticGrooming = "AutomaticGrooming",
+}
+
 export interface UserProjectData {
     org : string,
     name : string,